Editor/patrol handles (#50)
An Impromptu Editor Tool -> so no Jira Ticket associated

To make patrol design easier, I just created a small handle Tool to allow users to easily set the patrol paths (for objects with the patrol script attached) in the scene itself

![Sc1](https://user-images.githubusercontent.com/23490604/123168011-3fdc8c00-d42c-11eb-9f5c-9dff0d1db83b.png)

**Setting Patrol Points**
![Moving](https://user-images.githubusercontent.com/23490604/123168119-639fd200-d42c-11eb-87e3-8bd4c9c5fee3.gif)

**Adding New Patrol Points**

there are two ways to add new patrol points. Firstly, by click CTRL/CMD + Left Click on the scene will _append_ a new patrol point to list of current patrol points

the other way is to click SHIFT + Left Click on the scene, this will insert a patrol point between the two nearest patrol points to the location at which the mouse was clicked (for the case where you want a new point in between point A and point B) 
_**Note that this type of addition can sometimes be finicky if there are multiple line segments nearby to the mouse click**_

![Adding](https://user-images.githubusercontent.com/23490604/123168665-225bf200-d42d-11eb-9354-f3bd3373fb7c.gif)

**Deleting Patrol Points**

You can delete a patrol point by Right-clicking the position you would like to delete

![Deleting](https://user-images.githubusercontent.com/23490604/123168745-415a8400-d42d-11eb-97d8-f2c1823dcc54.gif)

**Note**
There is also Undo/Redo Fucntionality enabled but for some reason if you do 2+ additions/removals in a row, the Undo can be a bit wonky. This can easily be adjusted and doesn't break anything, but I will look into it in the future if needed