Gursaanj/LAZ-75 (#63)
**JIRA Ticket**
https://perigongames.atlassian.net/browse/LAZ-75

Added checkpoints within the Sandbox scene (as drawn in light pink spheres). The initial checkpoint is also the initial spawn position of Laz. Once killed Laz will respawn in the intended Checkpoint.

In build, there is no indicator as to if laz has entered a checkpoint. Will address this if needed

**Screenshots / GIFs**
![image](https://user-images.githubusercontent.com/23490604/126250967-ba7a928f-b6b6-4b07-b20a-d02fcd5083f5.png)